+++ R/count.R
@@ -11,10 +11,8 @@ setMethod(
   signature = c(x = "data.frame"),
   definition = function(x, f, margin = 1, negate = FALSE, na.rm = FALSE, ...) {
     assert_function(f)
-
     if (negate) f <- Negate(f)
-    x <- lapply(X = x, FUN = f, ...)
-    x <- do.call(cbind.data.frame, x)
+    x[] <- lapply(X = x, FUN = f, ...)
     if (any(margin == 1)) return(rowSums(x, na.rm = na.rm))
     if (any(margin == 2)) return(colSums(x, na.rm = na.rm))
     x