+++ R/keep.R
@@ -10,8 +10,10 @@ setMethod(
   f = "keep",
   signature = c(x = "ANY"),
   definition = function(x, f, margin = 1, negate = FALSE, all = FALSE,
-                        verbose = getOption("arkhe.verbose"), ...) {
-    i <- detect(x, f = f, margin = margin, negate = negate, all = all, ...)
+                        na.rm = FALSE, verbose = getOption("arkhe.verbose"),
+                        ...) {
+    i <- detect(x, f = f, margin = margin, negate = negate, all = all,
+                na.rm = na.rm, ...)
     discard_message(x, keep = i, margin = margin, verbose = verbose)
     if (any(margin == 1)) return(x[i, , drop = FALSE])
     if (any(margin == 2)) return(x[, i, drop = FALSE])
@@ -25,9 +27,10 @@ setMethod(
 setMethod(
   f = "keep_rows",
   signature = c(x = "ANY"),
-  definition = function(x, f, negate = FALSE, all = FALSE,
+  definition = function(x, f, negate = FALSE, all = FALSE, na.rm = FALSE,
                         verbose = getOption("arkhe.verbose"), ...) {
-    keep(x, f, margin = 1, negate = negate, all = all, verbose = verbose, ...)
+    keep(x, f, margin = 1, negate = negate, all = all, na.rm = na.rm,
+         verbose = verbose, ...)
   }
 )
 
@@ -37,8 +40,9 @@ setMethod(
 setMethod(
   f = "keep_cols",
   signature = c(x = "ANY"),
-  definition = function(x, f, negate = FALSE, all = FALSE,
+  definition = function(x, f, negate = FALSE, all = FALSE, na.rm = FALSE,
                         verbose = getOption("arkhe.verbose"), ...) {
-    keep(x, f, margin = 2, negate = negate, all = all, verbose = verbose, ...)
+    keep(x, f, margin = 2, negate = negate, all = all, na.rm = na.rm,
+         verbose = verbose, ...)
   }
 )