+++ inst/tinytest/test_compact.R
@@ -45,23 +45,29 @@ expect_equal(nrow(clean_col), nrow(charBlank))
 # Compact (logical) ============================================================
 bin <- matrix(c(FALSE, TRUE, TRUE, TRUE, FALSE,
                 FALSE, TRUE, FALSE, FALSE, TRUE,
-                FALSE, FALSE, FALSE, TRUE, FALSE),
+                FALSE, FALSE, FALSE, FALSE, FALSE),
               nrow = 5, ncol = 3)
 
 ## Nothing to remove
 clean <- compact_rows(bin)
-expect_equal(dim(clean), dim(bin))
+expect_equal(nrow(clean), 4)
+clean <- compact_cols(bin)
+expect_equal(ncol(clean), 2)
 
 binNA <- bin
-binNA[1, ] <- NA # Add NA
-binNA[, 1] <- NA # Add NA
+binNA[1, 1] <- NA # Add NA
+binNA[5, 3] <- NA # Add NA
 
 ## Remove rows
-clean_row <- compact_rows(binNA)
+clean_row <- compact_rows(binNA) # Nothing to remove
+expect_equal(dim(clean_row), dim(binNA))
+clean_row <- compact_rows(binNA, na.rm = TRUE)
 expect_equal(nrow(clean_row), 4)
 expect_equal(ncol(clean_row), ncol(binNA))
 
 ## Remove columns
-clean_col <- compact_cols(binNA)
+clean_col <- compact_cols(binNA) # Nothing to remove
+expect_equal(dim(clean_col), dim(binNA))
+clean_col <- compact_cols(binNA, na.rm = TRUE)
 expect_equal(ncol(clean_col), 2)
 expect_equal(nrow(clean_col), nrow(binNA))